refactoring tools list container
